% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-AddOrReplacePrecipitation.r
\name{addOrReplacePrecipitation}
\alias{DailyPrecipitationSerie-methods}
\alias{addOrReplacePrecipitation}
\alias{addOrReplacePrecipitation-method}
\title{Add or replace precipitation}
\usage{
addOrReplacePrecipitation(object, precipitation, date)
}
\arguments{
\item{object}{an object of class DailyPrecipitationSerie}

\item{precipitation}{the precipitation amount in millimeters}

\item{date}{the date of the precipitation amount accurance}
}
\value{
the \code{object} modified with the new precipitation value
}
\description{
It adds or replaces a precipitation value in millimetesr along with its occurance date into
an object of class DailyPrecipitationSerie.
}
\author{
Lucas Venezian Povoa
}

