% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-AddPrecipitation.r
\name{addPrecipitation}
\alias{DailyPrecipitationSerie-methods}
\alias{addPrecipitation}
\alias{addPrecipitation-method}
\title{Add precipitation}
\usage{
addPrecipitation(object, precipitation, date)
}
\arguments{
\item{object}{an object of class DailyPrecipitationSerie}

\item{precipitation}{the precipitation amount in millimeters}

\item{date}{the date of the precipitation amount accurance}
}
\value{
the modified object
}
\description{
Add a precipitation value in millimeters along with its respective date into
an object of class DailyPrecipitationSerie.
}
\author{
Lucas Venezian Povoa
}

