% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-as.data.frame.r
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\title{Convert to data.frame}
\usage{
\S4method{as.data.frame}{DailyPrecipitationSerie}(x)
}
\value{
A data.frame with 33 columns, where the first refers to year, the second refers to months,
and the remainder for precipitation in millimenters for each day.
}
\description{
It converts a DailyPrecipitationSerie into a data.frame.
}
\author{
Lucas Venezian Povoa
}

