% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Function-as.date.r
\name{as.date}
\alias{Date}
\alias{as.date}
\title{Date}
\usage{
Date(day = NA, month = NA, year = NA)
}
\arguments{
\item{day}{the day}

\item{month}{the month in number format}

\item{year}{the year with four digits}
}
\value{
a Date object
}
\description{
It creates an object of class Date
}
\author{
Lucas Venezian Povoa
}

