% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Function-complete.serie.r
\name{complete.serie}
\alias{DailyPrecipitationSerie-methods}
\alias{complete.serie}
\title{Completing Daily Precipitation Serie}
\usage{
complete.serie(serie, method)
}
\arguments{
\item{method}{what type of method should be used. Possible methods are:
\item{mlp}{for Multilayer Perceptron Artifical Neural Network}
\item{glm}{for Generalized Linear Regression}}

\item{object}{an object of the class DailyPrecipitationSerie}
}
\description{
It completes missing dates into a daily precipitation
serie using a statistical method for forecasting missing values.
}
\author{
Lucas Venezian Povoa
}

