% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Function-DailyPrecipitationSerie.r
\docType{function}
\name{DailyPrecipitationSerie}
\alias{DailyPrecipitationSerie}
\alias{DailyPrecipitationSerie-methods}
\title{DailyPrecipitationSerie Constructor}
\usage{
DailyPrecipitationSerie(precipitation = numeric(0), date = as.Date(character(0)), lat = 0, long = 0, elevation = 0, station_name = "")
}
\arguments{
\item{precipitation}{the precipitation values in millimeters of the serie}

\item{date}{the dates of the serie}

\item{lat}{the latitude position of the sensor}

\item{long}{the longitude position of the sensor}

\item{elevation}{the elevation of the sensor}

\item{station_name}{the name of the sensor}
}
\value{
an object of the class DailyPrecipitationSerie
}
\description{
It constructs an instance of the DailyPrecipitationSerie class.
}
\author{
Lucas Venezian Povoa
}

