% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-export.data.r
\name{export.data}
\alias{DailyPrecipitationSerie-methods}
\alias{export.data}
\title{Export data}
\usage{
export.data(serie, file, sep = ",", dec = ".")
}
\description{
It exports an object of class DailyPrecipitationSerie to a CSV file.
}
\author{
Lucas Venezian Povoa
}

