% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-length.r
\docType{methods}
\name{length}
\alias{DailyPrecipitationSerie-methods}
\alias{length}
\title{length}
\usage{
\S4method{length}{DailyPrecipitationSerie}(x)
}
\description{
It returns the number of observations into a daily precipitation serie.
}
\author{
Lucas Venezian Povoa
}

