% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-max.r
\name{max}
\alias{max}
\title{Max precipitation along with its date}
\usage{
max(x, ..., na.rm = FALSE)
}
\value{
the first maximum precipitation value along with its respective date.
}
\description{
It gets the first minimum precipitation value along with its respective date.
}
\author{
Lucas Venezian Povoa
}

