% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-min.r
\name{min}
\alias{min}
\title{Min precipitation along with its date}
\usage{
min(x, ..., na.rm = FALSE)
}
\value{
the first minimum precipitation value along with its respective date.
}
\description{
It gets the first minimum precipitation value along with its respective date.
}
\author{
Lucas Venezian Povoa
}

