% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-MissingDates.r
\name{missingDates}
\alias{DailyPrecipitationSerie-methods}
\alias{missingDates}
\title{Get Missing Dates}
\usage{
missing.dates(object)
}
\description{
It gets the missing dates into a precipitation serie described via an object of class
DailyPrecipitationSerie.
}

