% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Function-mlp4precipitation.r
\name{mlp}
\alias{mlp}
\alias{mlp4precipitation}
\title{Multilayer Perceptron (MLP)}
\usage{
mlp4precipitation(...)
}
\arguments{
\item{...}{a set of objects of DailyPrecipitationSerie}
}
\description{
A MLP for completing missing dates into a precipitation serie.
}

