% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Package-prefann.r
\docType{package}
\name{prefann}
\alias{prefann}
\alias{prefann-package}
\title{Precipitation forecasting with Artificial Neural Network and others}
\description{
It contains functions to forecast missing precipitation
values using a daily precipitation serie.
}
\author{
Lucas Venezian Povoa \email{lucasvenez@gmail.com}
}
\keyword{forecast}
\keyword{missing}
\keyword{precipitation}
\keyword{values}

