% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Function-read.data.r
\name{read.data}
\alias{DailyPrecipitationSerie-methods}
\alias{read.data}
\title{Read Precipitation Serie}
\usage{
read.data(file_path, sep = ",",
  quote = "\\"", dec = ".", header = TRUE, na.value = NA, ...)
}
\value{
An object of the class DailyPrecipitationSerie.
}
\description{
It reads a precipitation serie from a CSV file.
}
\details{
The CSV file should be structure with 33 columns, where each line represents a sample.
The 1st column contains the year, the 2nd column contains the month, and the columns 3-33 contain
the precipitation in milimmeters for each day in the month/year.
}
\author{
Lucas Venezian Povoa
}

