% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-ReplacePrecipitation.r
\name{replacePrecipitation}
\alias{DailyPrecipitationSerie-methods}
\alias{replacePrecipitation}
\alias{replacePrecipitation-method}
\title{Replace precipitation value}
\usage{
replacePrecipitation(object, precipitation, date)
}
\arguments{
\item{object}{an object of class DailyPrecipitationSerie}

\item{precipitation}{the precipitation amount in millimeters}

\item{date}{the date of the precipitation amount accurance}
}
\value{
the modified object
}
\description{
It replaces a precipitation value of a DailyPrecipitationSerie object.
}
\author{
Lucas Venezian Povoa
}

