% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Function-series.format.r
\name{series.format}
\alias{DailyPrecipitationSerie-methods}
\alias{series.format}
\title{Compact daily precipitation series}
\usage{
series.format(...)
}
\arguments{
\item{...}{a set of objects of class DailyPrecipitationSerie}
}
\value{
A data.frame with all input data
}
\description{
It compacts daily precipitation
series into a unique data.frame
}
\author{
Lucas Venezian Povoa
}

