% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Method-summary.r
\docType{methods}
\name{summary}
\alias{DailyPrecipitationSerie-methods}
\alias{summary}
\alias{summary.DailyPrecipitationSerie}
\title{DailyPrecipitationSerie summary}
\usage{
\S4method{summary}{DailyPrecipitationSerie}(object)
}
\description{
It summarizes an object of class DailyPrecipitationSerie.
}
\author{
Lucas Venezian Povoa
}

