% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class-DailyPrecipitationSerie.r, R/Function-DailyPrecipitationSerie.r
\docType{class}
\name{DailyPrecipitationSerie}
\alias{DailyPrecipitationSerie}
\alias{DailyPrecipitationSerie-class}
\alias{DailyPrecipitationSerie-methods}
\title{The DailyPrecipitationSerieClass}
\usage{
DailyPrecipitationSerie(precipitation = numeric(0), date = as.Date(character(0)), lat = 0, long = 0, elevation = 0, station_name = "")
}
\arguments{
\item{precipitation}{the precipitation values in millimeters of the serie}

\item{date}{the dates of the serie}

\item{lat}{the latitude position of the sensor}

\item{long}{the longitude position of the sensor}

\item{elevation}{the elevation of the sensor}

\item{station_name}{the name of the sensor}
}
\value{
an object of the class DailyPrecipitationSerie
}
\description{
This class encompasses a daily precipitation serie, validating its input against duplicate dates
and invalid precipitations.

It constructs an instance of the DailyPrecipitationSerie class.
}
\section{Slots}{

\description {
   \item{\code{precipitation}}{the precipitation values in millimeters of the serie}
   \item{\code{date}}{the dates of the serie}
   \item{\code{lat}}{the latitude position of the precipitation sensor}
   \item{\code{long}}{the longitude position of the precipitation sensor}
   \item{\code{elevation}}{the elvation of the precipitation sensor}
   \item{\code{station_name}}{the name of the precipitation sensor}
}
}
\author{
Lucas Venezian Povoa

Lucas Venezian Povoa
}

